Show failure cause for tests expected to fail
We have tests which are expected to fail, defined alongside the tests
that are meant to pass and this is only checked with a flag defined in
the test itself.

So when a test is meant to fail, we just see if the testable funtion is
returning error and have no idea if the error is the one we wanted to
see.

Hence add a log statement to see if test fails for the reason we want &
not something else.