BREAKING CHANGE(spec): Change volumeName to volumeRef
Right now we refer a volume name from mounts section of containers
with field called volumeName, change it to volumeRef.

This also introduces a convention of naming fields in OpenCompose
spec. So any field that is referencing to some other field it should
end with 'Ref'.

Breaks the way we refer the volume name.

Fixes #119