add validation for environment variables
This commit adds validation in object.go for environment variables.
Currently, the only validation that is being added is to check for
the presence of "=" in either environment variable key or value,
and if found, then error out.

This is only going to validate the environment variables passed in
the OpenCompose file, not as per Kubernetes validations, which
should be added in kubernetes.go which is being tracked in #139