Update reference documentation
This updates the reference documenation in order to have a clearer image
in terms of what's going on with OpenCompose.

We also increase the spacing in-between the titles and rename a few
specification names (ex. PortSpec to Port) for simplicity.

We also switch to example-based learning. Going through each part of the
specification full-example.