add a mandatory "name" field to define containers
This commit adds a mandatory field "name" to the containers
definition in the OpenCompose spec. Using this field, we can
mention the name of the container.

Prior to this commit, a minimal OpenCompose file would look like -

```
version: "0.1-dev"

services:
- name: foobar
  containers:
  - image: foo/bar:tag

```

Now, a "name" field for every container is mandatory to be
specified-

```
version: "0.1-dev"

services:
- name: foobar
  containers:
  - name: baz
    image: foo/bar:tag

```

This commit also fixes docs, updates tests for the changes,
updates examples.